requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-background-tasks==1.2.5
django-ckeditor==6.7.0
django-compat==1.0.15
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.1
django-extensions-shell==1.7.4.1
django-filter==23.2
django-js-asset==2.1.0
django-redis==5.3.0
django-shell==1.0.2
django-sitemap-generator==4.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dp==1.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
flake8==6.0.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kombu==5.3.1
magic-filter==1.0.11
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pipdeptree==2.13.0
pluggy==1.3.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.0.1
Pygments==2.15.1
pygount==1.6.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.5.5
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.13.1
sentry-sdk==1.31.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==5.4.0
telegram==0.0.1
transliterate==1.10.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
xlwt==1.3.0
yarl==1.9.2
